pyproject.toml
[tool.poetry]
name = "natije"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
pillow = "^10.2.0"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
pydantic-settings = "^2.1.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
bcrypt = "^4.1.2"
django-cors-headers = "^4.3.1"
opencv-python = "^4.9.0.80"
pytube = "^15.0.0"
aiogram = "^3.3.0"
django-silk = "^5.0.4"
celery = "^5.3.6"
adrf = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
